+++ include/diplib/geometry.h
@@ -43,37 +43,37 @@ namespace dip {
 /// called `interpolationMethod`, which determines how image data is interpolated. It can be set to one of
 /// the following strings:
 ///
-///  - '"3-cubic"' (or `""`): Third-order cubic spline interpolation (Keys, 1981), using 4 input samples to
+///  - `"3-cubic"` (or `""`): Third-order cubic spline interpolation (Keys, 1981), using 4 input samples to
 ///    compute each output sample. This is the default method for most functions.
 ///
-///  - '"4-cubic"': Fourth-order cubic spline interpolation (Keys, 1981), using 6 input samples to compute
+///  - `"4-cubic"`: Fourth-order cubic spline interpolation (Keys, 1981), using 6 input samples to compute
 ///    compute each output sample.
 ///
-///  - '"linear"': Linear interpolation, using 2 input samples to compute each output sample.
+///  - `"linear"`: Linear interpolation, using 2 input samples to compute each output sample.
 ///
-///  - '"nearest"' (or '"nn"'): Nearest neighbor interpolation, samples are simply shifted and replicated.
+///  - `"nearest"` (or `"nn"`): Nearest neighbor interpolation, samples are simply shifted and replicated.
 ///
-///  - '"inverse nearest"' (or '"nn2"'): Nearest neighbor interpolation, but resolves the rounding of x.5 in
+///  - `"inverse nearest"` (or `"nn2"`): Nearest neighbor interpolation, but resolves the rounding of x.5 in
 ///    the opposite direction that `"nearest"` does. This is useful when applying the inverse of an earlier
 ///    transform, to be able to obtain the original geometry back.
 ///
-///  - '"bspline"': A third-order cardinal B-spline is computed for all samples on an image line, which is
+///  - `"bspline"`: A third-order cardinal B-spline is computed for all samples on an image line, which is
 ///    sampled anew to obtain the interpolated sample values. All input samples are used to compute all
 ///    output samples, but only about 10 input samples significantly influence each output value.
 ///
-///  - '"lanczos8"': Lanczos interpolation with *a* = 8, using 16 input samples to compute each output sample.
+///  - `"lanczos8"`: Lanczos interpolation with *a* = 8, using 16 input samples to compute each output sample.
 ///    The Lanczos kernel is a sinc function windowed by a larger sinc function, where *a* is the width of
 ///    the larger sinc function. The kernel is normalized.
 ///
-///  - '"lanczos6"': Lanczos interpolation with *a* = 6, using 12 input samples to compute each output sample.
+///  - `"lanczos6"`: Lanczos interpolation with *a* = 6, using 12 input samples to compute each output sample.
 ///
-///  - '"lanczos4"': Lanczos interpolation with *a* = 4, using 8 input samples to compute each output sample.
+///  - `"lanczos4"`: Lanczos interpolation with *a* = 4, using 8 input samples to compute each output sample.
 ///
-///  - '"lanczos3"': Lanczos interpolation with *a* = 3, using 6 input samples to compute each output sample.
+///  - `"lanczos3"`: Lanczos interpolation with *a* = 3, using 6 input samples to compute each output sample.
 ///
-///  - '"lanczos2"': Lanczos interpolation with *a* = 2, using 4 input samples to compute each output sample.
+///  - `"lanczos2"`: Lanczos interpolation with *a* = 2, using 4 input samples to compute each output sample.
 ///
-///  - '"ft"': Interpolation through padding and cropping, and/or modifying the phase component of the
+///  - `"ft"`: Interpolation through padding and cropping, and/or modifying the phase component of the
 ///    Fourier transform of the image line. Padding with zeros increases the sampling density, cropping
 ///    reduces the sampling density, and multiplying the phase component by \f$-j s \omega\f$ shifts
 ///    the image. Equivalent to interpolation with a sinc kernel. All input samples are used to compute
@@ -205,6 +205,42 @@ inline Image Shift(
    Shift( in, out, shift, interpolationMethod, boundaryCondition );
    return out;
 }
+
+
+/// \brief Finds the values of the image at sub-pixel locations `coordinates` by linear interpolation.
+///
+/// The array `coordinates` must have all elements be array of the same length as the image dimensionality.
+/// Any coordinates outside of the image domain are returned as zero values. That is, no extrapolation is
+/// performed.
+///
+/// `interpolationMethod` has a restricted set of options: `"linear"`, `"3-cubic"`, or `"nearest"`.
+/// See \ref interpolation_methods for their definition.
+///
+/// `out` will be a 1D image with the same size as the `coordinates` array, and the same data type and tensor
+/// shape as `in`. To obtain results in a floating-point type, set the data type of `out` and protect it,
+/// see \ref protect.
+DIP_EXPORT void ResampleAt(
+      Image const& in,
+      Image& out,
+      FloatCoordinateArray const& coordinates,
+      String const& interpolationMethod = S::LINEAR
+);
+inline Image ResampleAt(
+      Image const& in,
+      FloatCoordinateArray const& coordinates,
+      String const& interpolationMethod = S::LINEAR
+) {
+   Image out;
+   ResampleAt( in, out, coordinates, interpolationMethod );
+   return out;
+}
+
+/// \brief Identical to the previous function with the same name, but for a single point.
+DIP_EXPORT Image::Pixel ResampleAt(
+      Image const& in,
+      FloatArray const& coordinates,
+      String const& interpolationMethod = S::LINEAR
+);
 
 
 // Undocumented internal function called by the other forms of Skew.